+++ domains/certificates/Query.API/API.AppTests/API.AppTests.csproj
@@ -29,4 +29,9 @@
     <ProjectReference Include="..\API\API.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Extensions\" />
+    <Folder Include="Infrastructure\" />
+  </ItemGroup>
+
 </Project>
