+++ domains/certificates/Query.API/API.AppTests/Infrastructure/MartenDbContainer.cs
@@ -1,0 +1,39 @@
+﻿using System.Threading.Tasks;
+using DotNet.Testcontainers.Builders;
+using DotNet.Testcontainers.Configurations;
+using DotNet.Testcontainers.Containers;
+using Xunit;
+
+namespace API.AppTests.Infrastructure;
+
+public class MartenDbContainer : IAsyncLifetime
+{
+    private readonly PostgreSqlTestcontainer testContainer;
+
+    public MartenDbContainer() =>
+        testContainer = new TestcontainersBuilder<PostgreSqlTestcontainer>()
+            //.WithCleanUp(true)
+            .WithDatabase(new PostgreSqlTestcontainerConfiguration
+            {
+                Database = "marten",
+                Username = "postgres",
+                Password = "postgres",
+            })
+            .WithImage("sibedge/postgres-plv8")
+            .Build();
+
+    public string ConnectionString => testContainer.ConnectionString;
+
+    public async Task InitializeAsync()
+    {
+        await testContainer.StartAsync();
+
+        await testContainer.ExecAsync(new[]
+        {
+            "/bin/sh", "-c",
+            "psql -U postgres -c \"CREATE EXTENSION plv8; SELECT extversion FROM pg_extension WHERE extname = 'plv8';\""
+        });
+    }
+
+    public Task DisposeAsync() => testContainer.DisposeAsync().AsTask();
+}
