+++ domains/certificates/Query.API/API.AppTests/Infrastructure/QueryApiWebApplicationFactory.cs
@@ -5,32 +5,54 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Security.Claims;
 using System.Text;
+using API.AppTests.Mocks;
 using API.DataSyncSyncer;
+using API.MasterDataService;
+using MassTransit;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.AspNetCore.TestHost;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.IdentityModel.Tokens;
 
-namespace API.AppTests;
+namespace API.AppTests.Infrastructure;
 
 public class QueryApiWebApplicationFactory : WebApplicationFactory<Program>
 {
-    protected override void ConfigureWebHost(IWebHostBuilder builder) =>
+    private readonly TestMasterDataService masterDataServiceMock;
+
+    public QueryApiWebApplicationFactory() => masterDataServiceMock = new TestMasterDataService();
+
+    public string MartenConnectionString { get; set; } = "";
+
+    protected override void ConfigureWebHost(IWebHostBuilder builder)
+    {
+        builder.UseSetting("ConnectionStrings:Marten", MartenConnectionString);
+
         builder.ConfigureTestServices(services =>
         {
             //Remove DataSyncSyncerWorker
             services.Remove(services.First(s => s.ImplementationType == typeof(DataSyncSyncerWorker)));
+
+            // Replace IMasterDataService
+            services.Remove(services.First(s => s.ServiceType == typeof(IMasterDataService)));
+            services.AddSingleton<IMasterDataService>(masterDataServiceMock);
         });
+    }
 
     public HttpClient CreateUnauthenticatedClient() => CreateClient();
 
-    public HttpClient CreateAuthenticatedClient()
+    public HttpClient CreateAuthenticatedClient(string subject)
     {
         var client = CreateClient();
-        client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", GenerateToken());
+        client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", GenerateToken(subject: subject));
 
         return client;
     }
+
+    public IBus GetMassTransitBus() => Services.GetRequiredService<IBus>();
+
+    public void AddMasterData(MasterData data) => masterDataServiceMock.Add(data);
 
     private static string GenerateToken(
         string scope = "",
